Updated the TODO list
This project has only begun, so I am going to need to set a todo list so I can keep myself on track. In the future, I may have an actually viable program. For now, though, it is merely a final project for a computer's class, so it isn't much at all