More hopes and dreams
After using this in a practical use, I figured out that there were quite a few holes in this code. Hence, the TODO needs an update. As well as this, I made a separate branch for TODOs in the very far flung scenario that there are other people working on the project who may get annoyed at such updates not really changing anything.